Add admin reservation management features and update response messages
- Implement update reservation endpoint with validation and status checks.
- Add endpoints for retrieving reservations by ID and paginated user reservations.
- Introduce error messages for reservation not found and status change restrictions.
- Enhance ReservationService with methods for updating and retrieving reservations.
- Create ReservationUpdateRequest DTO for handling update requests.
- Update ResponseMessage with new reservation-related messages.
- Modify UserService and CarService to include checks for existing reservations.

+++ src/main/java/com/visionrent/service/CarService.java
@@ -32,6 +32,9 @@ public class CarService {
     @Autowired
     private CarMapper carMapper;
 
+    @Autowired
+    private ReservationService reservationService;
+
     public void saveCar(UUID imageId, CarDTO carDTO) {
 
         // image Id image repo da var mı?
@@ -124,6 +127,11 @@ public class CarService {
 
         if (car.getBuiltIn()){
             throw new BadRequestException(ErrorMessage.NOT_PERMITTED_METHOD_MESSAGE);
+        }
+
+        boolean exist = reservationService.existsByCar(car);
+        if (exist){
+            throw new BadRequestException(ErrorMessage.CAR_USED_BY_RESERVATION_MESSAGE);
         }
 
         carRepository.delete(car);
